feat: add full support for aws access keys
Closes: #1669
