+++ src/modules/aws_iam/entity/index.ts
@@ -1,2 +1,3 @@
 export * from './role';
 export * from './user';
+export * from './access_key';
